Add the xmlns and xmlns:xlink attributes
With the correct namespace URI for each as a single-option enum.

I've only added the attributes to the content model for `<svg>`, although of course the attributes are really global.
But it is normal with SVG to only declare the namespaces on the root element. If you were going to use `xmlns` anywhere else, it would probably be to set a different namespace URI on child elements of `<metadata>` or `<foreignObject>`.  But I wasn't sure how to add that option to your code!